# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Pet Shop",
        "description": "Pet shop API documentation",
        "contact": {
            "name": "Rabiul Islam",
            "email": "rir.cse.71@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
        "/product/create": {
            "post": {
                "description": "Create a new product",
                "operationId": "7d084d79a3a2164eb9ed2d675612dfc2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_uuid",
                                    "price",
                                    "metadata",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "New product"
                                    },
                                    "category_uuid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "100"
                                    },
                                    "metadata[file_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "metadata[brand_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Lorem ipsum.."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "400": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/product/{uuid}": {
            "get": {
                "description": "Get single product",
                "operationId": "6eb40cc9091f67781fe34704490c5c0b",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product retrieved successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "description": "Update product",
                "operationId": "9ee612579b1aeb5b1827cbeeb96bfd87",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category_uuid",
                                    "price",
                                    "metadata",
                                    "description"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "New product"
                                    },
                                    "category_uuid": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "format": "integer",
                                        "example": "100"
                                    },
                                    "metadata[file_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "metadata[brand_uuid]": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "1234-5678"
                                    },
                                    "description": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Lorem ipsum.."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "description": "Delete product",
                "operationId": "15d7fa8465a9f928f00d7e145da748b2",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID of product",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Product deleted successfully"
                    },
                    "400": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    }
}